Pre-selectivity corrected dataset consisting of pollock biomass & numbers at age on a per-interval basis. Winter GOA.

Pre-selectivity corrected dataset consisting of pollock biomass & numbers at age on a per-interval basis. Winter GOA.

PreSelectivityWinterGOAages

Format

A data frame with 236399 rows and 16 columns:

Ship

ship number; 21 is MF, 157 is DY (numeric)

Survey

survey number (numeric)

DataSetID

Duh (numeric)

AnalysisId

Duh (numeric)

ReportNumber

The report number associated with the ages (numeric)

Interval

Original interval number, not unique in older surveys (numeric)

Age

Age of fish (years), determined by age & growth

Transect

Original transect designation; NULL transects were removed from these data

Description

The original descriptor in old MB

Numbers

Numbers of fish in that edsu

Biomasskg

Biomass (kg) of fish in that edsu

START_TIME

Date and Time (UTC) at the start of the interval

Region

Region in which the survey was performed, used st_join(st_within) from the sf library and updated geographic regions stored in: 'G://ARCVIEW//exported_shapefiles//winter_survey_regions_BarnabasChiniak_NAD1983' to determine survey area

Latitude

Latitude at start of edsu; older surveys don't have an end lat/long

Longitude

Longitude at start of edsu; older surveys don't have an end lat/long

Year

Year of Survey

#' @source Historic MACE reported values